// src/types/API.ts
export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items:  Array<Comment | null >,
  nextToken?: string | null,
};
// ...
export type ModelSubscriptionCommentFilterInput = {
  and?: Array< ModelSubscriptionCommentFilterInput | null > | null,
  comment?: ModelSubscriptionStringInput | null,
  createdAt?: ModelSubscriptionStringInput | null,
  id?: ModelSubscriptionIDInput | null,
  or?: Array< ModelSubscriptionCommentFilterInput | null > | null,
  owner?: ModelStringInput | null,
  postId?: ModelSubscriptionIDInput | null,
  updatedAt?: ModelSubscriptionStringInput | null,
  userId?: ModelSubscriptionIDInput | null,
};
// ...
export type OnCreateCommentSubscriptionVariables = {
  filter?: ModelSubscriptionCommentFilterInput | null,
  owner?: string | null,
};

export type OnCreateCommentSubscription = {
  onCreateComment?:  {
    __typename: "Comment",
    comment: string,
    createdAt: string,
    id: string,
    like?:  {
      __typename: "Likes",
      commentId?: string | null,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type OnCreateLikesSubscriptionVariables = {
  filter?: ModelSubscriptionLikesFilterInput | null,
  owner?: string | null,
};

export type OnCreateLikesSubscription = {
  onCreateLikes?:  {
    __typename: "Likes",
    comment?:  {
      __typename: "Comment",
      comment: string,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    commentId?: string | null,
    createdAt: string,
    id: string,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type OnCreatePostSubscriptionVariables = {
  filter?: ModelSubscriptionPostFilterInput | null,
  owner?: string | null,
};

export type OnCreatePostSubscription = {
  onCreatePost?:  {
    __typename: "Post",
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    description?: string | null,
    id: string,
    image?: string | null,
    images: Array< string | null >,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    nofComments: number,
    nofLikes: number,
    owner?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
    video?: string | null,
  } | null,
};

export type OnCreateUserSubscriptionVariables = {
  filter?: ModelSubscriptionUserFilterInput | null,
  owner?: string | null,
};

export type OnCreateUserSubscription = {
  onCreateUser?:  {
    __typename: "User",
    bio?: string | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    email: string,
    id: string,
    image?: string | null,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    name: string,
    nofFollowers: number,
    nofFollowings: number,
    nofPosts: number,
    owner?: string | null,
    posts?:  {
      __typename: "ModelPostConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    username: string,
    website?: string | null,
  } | null,
};

export type OnDeleteCommentSubscriptionVariables = {
  filter?: ModelSubscriptionCommentFilterInput | null,
  owner?: string | null,
};

export type OnDeleteCommentSubscription = {
  onDeleteComment?:  {
    __typename: "Comment",
    comment: string,
    createdAt: string,
    id: string,
    like?:  {
      __typename: "Likes",
      commentId?: string | null,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type OnDeleteLikesSubscriptionVariables = {
  filter?: ModelSubscriptionLikesFilterInput | null,
  owner?: string | null,
};

export type OnDeleteLikesSubscription = {
  onDeleteLikes?:  {
    __typename: "Likes",
    comment?:  {
      __typename: "Comment",
      comment: string,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    commentId?: string | null,
    createdAt: string,
    id: string,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type OnDeletePostSubscriptionVariables = {
  filter?: ModelSubscriptionPostFilterInput | null,
  owner?: string | null,
};

export type OnDeletePostSubscription = {
  onDeletePost?:  {
    __typename: "Post",
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    description?: string | null,
    id: string,
    image?: string | null,
    images: Array< string | null >,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    nofComments: number,
    nofLikes: number,
    owner?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
    video?: string | null,
  } | null,
};

export type OnDeleteUserSubscriptionVariables = {
  filter?: ModelSubscriptionUserFilterInput | null,
  owner?: string | null,
};

export type OnDeleteUserSubscription = {
  onDeleteUser?:  {
    __typename: "User",
    bio?: string | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    email: string,
    id: string,
    image?: string | null,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    name: string,
    nofFollowers: number,
    nofFollowings: number,
    nofPosts: number,
    owner?: string | null,
    posts?:  {
      __typename: "ModelPostConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    username: string,
    website?: string | null,
  } | null,
};

export type OnUpdateCommentSubscriptionVariables = {
  filter?: ModelSubscriptionCommentFilterInput | null,
  owner?: string | null,
};

export type OnUpdateCommentSubscription = {
  onUpdateComment?:  {
    __typename: "Comment",
    comment: string,
    createdAt: string,
    id: string,
    like?:  {
      __typename: "Likes",
      commentId?: string | null,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type OnUpdateLikesSubscriptionVariables = {
  filter?: ModelSubscriptionLikesFilterInput | null,
  owner?: string | null,
};
// ...
export type OnUpdatePostSubscriptionVariables = {
  filter?: ModelSubscriptionPostFilterInput | null,
  owner?: string | null,
};
// ...
export type OnUpdateUserSubscriptionVariables = {
  filter?: ModelSubscriptionUserFilterInput | null,
  owner?: string | null,
};